Fix TOML parsing with large unsigned integers
The parser parses everything as signed integers so the configuration
file will fail to parse large numbers that don't fit in an i64.

https://github.com/alexcrichton/toml-rs/issues/256

+++ tools/yabridgectl/src/utils.rs
@@ -183,7 +183,7 @@ pub fn verify_wine_setup(config: &mut Config) -> Result<()> {
             yabridge_host_exe_so_path.display()
         )
     })?);
-    let yabridge_host_hash = hasher.finish();
+    let yabridge_host_hash = hasher.finish() as i64;
 
     // Since these checks can take over a second if wineserver isn't already running we'll only
     // perform them when something has changed
